refactor(cli): split main.rs into a clap surface and a dispatcher
main.rs was 1,085 lines holding the whole clap declaration and the
dispatch for every subcommand, which made it the file every unrelated
change had to touch. The declarations move to cli.rs (with the parse
tests) and the path/bootstrap/tracing helpers to startup.rs, leaving
main.rs at 510 lines of dispatch.
