+++ crates/outl-cli/src/main.rs
@@ -13,408 +13,24 @@
 //!   that lets Claude Desktop reach the same handlers over stdio.
 
 use anyhow::{Context, Result};
-use clap::{Parser, Subcommand};
-use std::path::{Path, PathBuf};
+use clap::Parser;
 
+mod cli;
 mod cmd;
 mod human;
 mod mcp;
 mod output;
+mod startup;
 mod sync_engine;
 mod workspace_layout;
 mod ws;
 
-#[derive(Parser, Debug)]
-#[command(
-    name = "outl",
-    about = "Local-first outliner with markdown as source of truth.",
-    long_about = "Local-first outliner with markdown as source of truth.\n\
-                  \n\
-                  Running `outl` with no subcommand opens the TUI in the workspace at \
-                  `--workspace` (default: current directory).",
-    version
-)]
-struct Cli {
-    /// Workspace path. Used by every subcommand that needs one;
-    /// defaults to the current directory. Subcommand-level positional
-    /// path, when provided, takes precedence.
-    #[arg(short = 'w', long, global = true, value_name = "DIR")]
-    workspace: Option<PathBuf>,
-
-    /// TUI theme preset (default-dark, light, logseq-light, dracula,
-    /// solarized-dark, nord, monokai, gruvbox). Overrides `[theme]
-    /// preset` in workspace `config.toml` for this run.
-    #[arg(long, global = true, value_name = "PRESET")]
-    theme: Option<String>,
-
-    #[command(subcommand)]
-    command: Option<Command>,
-
-    /// Increase verbosity. Pass multiple times for more detail.
-    #[arg(short, long, action = clap::ArgAction::Count, global = true)]
-    verbose: u8,
-}
-
-/// Subcommands for peer/device management.
-#[derive(Debug, clap::Subcommand)]
-enum PeerCommand {
-    /// Pair with another device. Prints a ticket (QR + string); run on both devices.
-    Pair {
-        /// Accept a ticket from the other device instead of generating one.
-        ///
-        /// `-` reads the ticket from stdin, which is how a 400-to-750-character
-        /// ticket gets into a container without going through argv:
-        /// `pbpaste | docker compose run -i --rm outl peer pair --ticket -`.
-        #[arg(long)]
-        ticket: Option<String>,
-        /// Human-readable name this device advertises to the other (shown in
-        /// its `peer list`). Defaults to the machine hostname.
-        #[arg(long)]
-        name: Option<String>,
-    },
-    /// Render a pairing ticket you already have as a scannable QR code.
-    ///
-    /// `outl peer pair` prints one itself, and skips it when the terminal is
-    /// too narrow for it to be readable. This is the way to get it back:
-    /// widen a window, or pipe the ticket in from wherever you copied it.
-    /// The phone app pairs by camera, so for a phone the QR is the only
-    /// practical route.
-    Qr {
-        /// The ticket. Omit it, or pass `-`, to read stdin.
-        ticket: Option<String>,
-    },
-    /// List all paired devices.
-    List,
-    /// Unpair a device by node-id prefix.
-    Remove {
-        /// Node-id prefix of the device to remove.
-        id: String,
-    },
-    /// Show connection status of all paired devices.
-    Status,
-    /// Lock out every paired device by rotating this workspace's identity.
-    ///
-    /// For a lost or stolen device. `peer remove` only takes effect on the
-    /// machine you run it on; this changes the workspace identity itself, so
-    /// no device that is not re-paired can sync again — including one you no
-    /// longer control.
-    ///
-    /// You will have to re-pair every device you still have.
-    RevokeAll {
-        /// Skip the confirmation prompt.
-        #[arg(long)]
-        yes: bool,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-enum Command {
-    /// Open the TUI on the workspace (default: `--workspace` or current dir).
-    Tui {
-        /// Workspace path. Overrides the global `--workspace`.
-        path: Option<PathBuf>,
-    },
-    /// Initialize a new workspace at the given path.
-    Init {
-        /// Workspace path. Created if it does not exist. Overrides `--workspace`.
-        path: Option<PathBuf>,
-        /// Op-log layout: `global` (single file per actor, legacy) or
-        /// `per-page` (one file per (actor, page) — Phase B of RFC #137).
-        /// New workspaces default to `global` for back-compat.
-        #[arg(long, default_value = "global", value_parser = ["global", "per-page"])]
-        scope: String,
-        /// Create the layout and config but write no ops: no
-        /// `templates/journal` page, no journal for today.
-        ///
-        /// For a workspace that exists only to become a replica of an
-        /// existing graph — a self-hosted `outl serve` box — and will
-        /// join it with `outl peer pair --ticket`. Pairing adopts the
-        /// host's workspace id but keeps this device's ops, so a seeded
-        /// replica pushes a second `templates/journal` page into the
-        /// host's graph. Do NOT use it for a workspace you intend to
-        /// write notes in directly; the journal template is missing.
-        #[arg(long)]
-        bare: bool,
-    },
-    /// Migrate a workspace's op log from `Global` (single file per
-    /// actor) to `PerPage` (one file per actor + page). RFC #137
-    /// Phase B. Reversible — the legacy file is preserved as
-    /// `ops-<actor>.jsonl.v0.bak`.
-    MigrateToPerPageOps {
-        /// Workspace path. Overrides `--workspace`.
-        path: Option<PathBuf>,
-    },
-    /// Run the background daemon: watch `.md` files and hold this device's
-    /// P2P endpoint so paired peers sync continuously.
-    ///
-    /// Both halves are on by default. The watcher reconciles external `.md`
-    /// edits into the op log; the sync half holds the iroh endpoint, deferring
-    /// to a GUI or TUI that already has it and taking over when that exits.
-    ///
-    /// `--no-watch` is the cheaper mode to leave running: without the watcher
-    /// there is nothing to write, so it takes no per-actor write lock and
-    /// cannot push a GUI onto a fresh ephemeral actor (and a fresh
-    /// `ops-<ulid>.jsonl`) on every launch. It still takes the P2P endpoint,
-    /// though, and does not give it back, so a GUI opened after this daemon
-    /// syncs through the shared ops/ dir until the daemon stops.
-    Serve {
-        /// Workspace path. Overrides the global `--workspace`.
-        path: Option<PathBuf>,
-        /// Reconcile every `.md` once and exit (no file watcher, no sync).
-        #[arg(long, conflicts_with_all = ["no_watch", "no_sync"])]
-        once: bool,
-        /// Skip the file watcher; only hold the P2P endpoint.
-        ///
-        /// Fails when `[sync] transport` is "file": with no watcher and no
-        /// endpoint to hold there is nothing left to do, and exiting 0 would
-        /// have a process manager restart it into that config forever.
-        #[arg(long)]
-        no_watch: bool,
-        /// Skip P2P sync; only run the file watcher.
-        #[arg(long)]
-        no_sync: bool,
-    },
-    /// Check workspace integrity.
-    Doctor {
-        /// Workspace path. Overrides the global `--workspace`.
-        path: Option<PathBuf>,
-        /// Emit the report as the JSON envelope instead of a human view.
-        #[arg(long)]
-        json: bool,
-        /// Apply the safe, reversible fixes the report lists as
-        /// repairable: re-project a stale `.md` from the op log,
-        /// rebuild a missing sidecar, drop a corrupt snapshot.
-        ///
-        /// Never touches `ops/`, never deletes a `.md`, never moves a
-        /// block to the trash. Every file it writes is copied to
-        /// `.outl/repair-backup/<timestamp>/` first.
-        #[arg(long)]
-        repair: bool,
-        /// Authorise a `--repair` whose measured volume is past the
-        /// point it runs unattended.
-        ///
-        /// The report always states how many content lines the page
-        /// re-projections would remove, and from how many pages, before
-        /// anything is written. Past the ceiling those writes stand
-        /// down; this is how you say "yes, I read that and I meant it".
-        #[arg(long, requires = "repair")]
-        force: bool,
-    },
-    /// Resolve orphan matches via the TUI.
-    Reconcile {
-        /// Workspace path. Overrides the global `--workspace`.
-        path: Option<PathBuf>,
-        /// Reconcile the pages whose `.md` holds content that exists in
-        /// no op, ignoring the sidecar hash gate.
-        ///
-        /// A page can be hash-faithful (sidecar agrees with the bytes on
-        /// disk) and still carry content the op log never saw, so the
-        /// ordinary reconcile skips it as in-sync — see issue #210. This
-        /// writes ops for that content, which is why it is opt-in.
-        ///
-        /// Run it only on a build whose parser preserves the content:
-        /// reconciling with a parser that drops prose after a block
-        /// property writes the truncated text into the log, making the
-        /// loss permanent.
-        #[arg(long = "ahead-of-log")]
-        ahead_of_log: bool,
-        /// Apply a deletion the orphan-volume guard refused.
-        ///
-        /// A reconcile that would trash more than 500 blocks of a page,
-        /// or more than 75% of one, stops and writes nothing — a `.md`
-        /// that arrived truncated (an undownloaded iCloud placeholder, a
-        /// half-flushed write) is indistinguishable from a real bulk
-        /// delete by shape, only by scale.
-        ///
-        /// This is the way to say the deletion was intended. Check what
-        /// the `.md` actually holds before reaching for it: the guard
-        /// fires on the case where the file is the thing that is wrong.
-        #[arg(long = "allow-bulk-delete")]
-        allow_bulk_delete: bool,
-    },
-    /// Recover block text that an `Op::Edit` truncated.
-    ///
-    /// The mirror of `reconcile --ahead-of-log`, reading the other
-    /// source. That one reads the `.md` and can only recover content
-    /// still on disk; this reads the **op log**, where a truncating
-    /// edit's predecessor still carries the full text — the only route
-    /// left for a page whose `.md` was already overwritten (issue #210).
-    ///
-    /// Read-only unless `--apply`. A restore is a new op; the op log is
-    /// never rewritten.
-    Recover {
-        /// Workspace path. Overrides the global `--workspace`.
-        path: Option<PathBuf>,
-        /// Write the recovered text back as new `Op::Edit`s.
-        ///
-        /// Additive by construction: a block only qualifies when its
-        /// current text is a prefix of the revision being restored, so
-        /// nothing it shows today is dropped.
-        #[arg(long)]
-        apply: bool,
-        /// Only report blocks that lost at least this many non-blank
-        /// lines. Raise it when the listing is too long to read; a
-        /// one-line loss is often ordinary editing.
-        #[arg(long = "min-lines", default_value_t = cmd::recover::DEFAULT_MIN_LINES, value_parser = clap::value_parser!(u16).range(1..))]
-        min_lines: u16,
-    },
-    /// Take, list, and restore local snapshots of the workspace.
-    ///
-    /// Uses the global `--workspace` for the target, like every other
-    /// subcommand-carrying command (a positional path would be
-    /// ambiguous against the subcommand name).
-    Backup {
-        #[command(subcommand)]
-        sub: cmd::backup::BackupSubcommand,
-    },
-    /// Inspect or list theme presets.
-    Theme {
-        #[command(subcommand)]
-        sub: Option<ThemeSubcommand>,
-    },
-    /// Import a graph from another outliner.
-    Import {
-        /// Source format: `roam` (JSON file), `logseq` (graph
-        /// directory), `obsidian` (vault directory), or `auto`
-        /// (detect from the source's shape).
-        format: String,
-        /// Path to the Logseq graph directory, the Roam backup file,
-        /// or the Obsidian vault directory.
-        src: PathBuf,
-        /// Destination workspace. Created if it doesn't exist yet.
-        dst: PathBuf,
-        /// Parse and report only — write nothing to the destination.
-        #[arg(long)]
-        dry_run: bool,
-        /// Print the import report as JSON.
-        #[arg(long)]
-        json: bool,
-        /// Keep source create/edit timestamps as `created::` /
-        /// `edited::` block properties.
-        #[arg(long)]
-        preserve_timestamps: bool,
-        /// Don't pull referenced files into `assets/` — keep the
-        /// original relative/remote links verbatim.
-        #[arg(long)]
-        no_assets: bool,
-        /// Import even when the destination already holds content —
-        /// overwrites those pages and discards anything written in outl,
-        /// or received from a paired device, since the last import.
-        #[arg(long)]
-        force: bool,
-    },
-    /// Import a file (PDF, image, …) and link it into the workspace.
-    Asset {
-        #[command(subcommand)]
-        sub: cmd::asset::AssetCommand,
-    },
-    /// Page-level operations.
-    Page {
-        #[command(subcommand)]
-        sub: cmd::page::PageCommand,
-    },
-    /// Manage workspace plugins (list / install / run / enable / disable).
-    Plugin {
-        #[command(subcommand)]
-        sub: cmd::plugin::PluginCommand,
-    },
-    /// Block-level operations.
-    Block {
-        #[command(subcommand)]
-        sub: cmd::block::BlockCommand,
-    },
-    /// Daily journal operations.
-    Daily {
-        #[command(subcommand)]
-        sub: cmd::daily::DailyCommand,
-    },
-    /// Full-text search.
-    Search(cmd::search::SearchArgs),
-    /// Structured query over pages.
-    Query(cmd::query::QueryArgs),
-    /// Backlinks and reference lookups.
-    Backlinks {
-        #[command(subcommand)]
-        sub: cmd::backlinks::BacklinksCommand,
-    },
-    /// Apply a list of write ops sequentially in one workspace session.
-    /// Reads `{"ops": [...]}` from stdin by default.
-    Batch(cmd::batch::BatchArgs),
-    /// Tag listing and lookups.
-    Tag {
-        #[command(subcommand)]
-        sub: cmd::tag::TagCommand,
-    },
-    /// Template operations (list, apply, resolve callable).
-    Template {
-        #[command(subcommand)]
-        sub: cmd::template::TemplateCommand,
-    },
-    /// Render a page in a target format (hugo / md / json).
-    Export {
-        #[command(subcommand)]
-        sub: Option<cmd::export_v2::ExportCommand>,
-        /// Legacy placeholder for `--to <fmt>` shape; only `hugo` was
-        /// ever accepted. Kept so prior scripts don't break.
-        #[arg(long)]
-        to: Option<String>,
-    },
-    /// Workspace summary (path, actor, counts).
-    Workspace {
-        #[command(subcommand)]
-        sub: WorkspaceSubcommand,
-    },
-    /// Run the MCP (Model Context Protocol) server over stdio. Wire
-    /// this into `claude_desktop_config.json` to expose every CLI
-    /// subcommand as an MCP tool.
-    Mcp {
-        #[command(subcommand)]
-        sub: McpSubcommand,
-    },
-    /// Manage peer devices for P2P sync.
-    Peer {
-        #[command(subcommand)]
-        cmd: PeerCommand,
-    },
-    /// Force a one-shot P2P sync pass against every paired device, then exit.
-    ///
-    /// For scripts that mutate via the CLI and must flush to peers before the
-    /// process dies — a normal `outl page/block/...` command is too short-lived
-    /// to bind an iroh endpoint, so it relies on whichever long-lived process
-    /// on this device holds the endpoint (a GUI, a TUI, `outl mcp serve`, or
-    /// `outl serve`) plus the catch-up re-sync instead. `outl sync` is the explicit flush; if one of
-    /// those already holds the endpoint it says so and exits, since that
-    /// process is already pushing these ops out.
-    ///
-    /// Exit codes: 0 a flush ran; 3 nothing was flushed (endpoint held
-    /// elsewhere, P2P off, or no paired device) so the ops are still local
-    /// until another process converges them; 1/2 the command failed.
-    Sync,
-}
-
-#[derive(Subcommand, Debug)]
-pub enum ThemeSubcommand {
-    /// Print every available preset, one per line.
-    List,
-    /// Describe a specific preset (palette + style names).
-    Show {
-        /// Preset name (case- and separator-insensitive).
-        name: String,
-    },
-}
-
-#[derive(Subcommand, Debug)]
-pub enum WorkspaceSubcommand {
-    /// Workspace info — path, actor, counts.
-    Info(cmd::workspace_info::WorkspaceInfoArgs),
-}
-
-#[derive(Subcommand, Debug)]
-pub enum McpSubcommand {
-    /// Start the MCP stdio server. Targets the workspace at the global
-    /// `--workspace` (or current directory if unset).
-    Serve {},
-}
+pub(crate) use cli::{
+    Cli, Command, McpSubcommand, PeerCommand, ThemeSubcommand, WorkspaceSubcommand,
+};
+use startup::{
+    default_device_name, ensure_workspace_or_prompt, init_tracing, resolve_init_path, resolve_path,
+};
 
 fn main() -> Result<()> {
     let cli = Cli::parse();
@@ -454,6 +70,15 @@ fn main() -> Result<()> {
         Some(Command::MigrateToPerPageOps { path }) => {
             let p = resolve_path(cli.workspace.as_ref(), path.as_ref())?;
             cmd::migrate_to_per_page_ops::run(&p)
+        }
+        Some(Command::Compact {
+            path,
+            apply,
+            no_horizon,
+            force,
+        }) => {
+            let p = resolve_path(cli.workspace.as_ref(), path.as_ref())?;
+            cmd::compact::run(&p, apply, no_horizon, force)
         }
         Some(Command::Serve {
             path,
@@ -883,165 +508,4 @@ fn run_sync(path: &std::path::Path) -> anyhow::Result<i32> {
         health.len()
     );
     Ok(output::EXIT_OK)
-}
-
-/// Resolve which workspace path to operate on.
-///
-/// Precedence (first hit wins):
-///
-/// 1. **Subcommand-positional** path (`outl page get … <path>`).
-/// 2. **Global `--workspace <DIR>`** flag.
-/// 3. **`workspace.last`** from `~/.config/outl/config.toml`
-///    (the same file the desktop's Settings modal writes when the
-///    user picks a workspace — so `outl` with no args lands on the
-///    workspace the user last opened in the GUI, no `--workspace`
-///    flag needed).
-/// 4. **Current directory** — final fallback (matches the
-///    `cd ~/notes && outl` muscle memory).
-///
-/// A path stored in `config.toml` that no longer exists on disk is
-/// skipped silently rather than failing the launch — the user
-/// likely deleted / unmounted the folder and would be surprised by
-/// a crash. The cwd fallback picks up.
-/// Best-effort device label for `outl peer pair` when `--name` is omitted.
-///
-/// Shells out to the `hostname` command (present on macOS + Linux) so the
-/// peer's device list reads "macbook" instead of a node-id stub, trimming the
-/// macOS `.local` suffix. Returns `None` if the command is unavailable or
-/// empty — pairing then advertises no alias, exactly as before this flag.
-/// Kept dependency-free on purpose; `--name` is the explicit override.
-fn default_device_name() -> Option<String> {
-    let out = std::process::Command::new("hostname").output().ok()?;
-    let raw = String::from_utf8(out.stdout).ok()?;
-    let name = raw.trim();
-    let name = name.strip_suffix(".local").unwrap_or(name);
-    (!name.is_empty()).then(|| name.to_string())
-}
-
-fn resolve_path(global: Option<&PathBuf>, local: Option<&PathBuf>) -> Result<PathBuf> {
-    if let Some(p) = local {
-        return Ok(p.clone());
-    }
-    if let Some(p) = global {
-        return Ok(p.clone());
-    }
-    if let Some(p) = outl_config::load().workspace.last {
-        if p.exists() {
-            return Ok(p);
-        }
-        tracing::warn!(
-            "config.toml workspace.last = {} is no longer on disk; falling back to cwd",
-            p.display()
-        );
-    }
-    std::env::current_dir().with_context(|| "reading current directory")
-}
-
-/// If `path` has no `.outl/` directory yet, prompt the user for
-/// permission to initialize one. If they say no, error out cleanly.
-///
-/// When stdin isn't a TTY (e.g. piped) we don't prompt — instead we
-/// error with the same "run `outl init`" message we used before. This
-/// keeps scripted callers predictable.
-fn ensure_workspace_or_prompt(path: &Path) -> Result<()> {
-    let outl_dir = path.join(".outl");
-    if outl_dir.exists() {
-        return Ok(());
-    }
-
-    use std::io::IsTerminal;
-    let interactive = std::io::stdin().is_terminal() && std::io::stderr().is_terminal();
-    if !interactive {
-        anyhow::bail!(
-            "no outl workspace at {} — run `outl init {}` first",
-            path.display(),
-            path.display()
-        );
-    }
-
-    use std::io::{BufRead, Write};
-    eprintln!("No outl workspace at {}.", path.display());
-    eprint!("Initialize a new workspace here? [y/N] ");
-    let _ = std::io::stderr().flush();
-    let mut line = String::new();
-    std::io::stdin()
-        .lock()
-        .read_line(&mut line)
-        .with_context(|| "reading prompt response")?;
-    let answer = line.trim().to_lowercase();
-    if answer == "y" || answer == "yes" {
-        cmd::init::run(path, "global", false)?;
-        Ok(())
-    } else {
-        anyhow::bail!("aborted — no workspace initialized at {}", path.display());
-    }
-}
-
-/// Same as [`resolve_path`] but errors out when neither flag nor positional
-/// was given (init refuses to create a workspace at the cwd by accident).
-fn resolve_init_path(global: Option<&PathBuf>, local: Option<&PathBuf>) -> Result<PathBuf> {
-    match local.or(global) {
-        Some(p) => Ok(p.clone()),
-        None => Err(anyhow::anyhow!(
-            "`outl init` needs an explicit path: pass a positional argument or `--workspace <DIR>`"
-        )),
-    }
-}
-
-fn init_tracing(verbosity: u8) {
-    let level = match verbosity {
-        0 => tracing::Level::INFO,
-        1 => tracing::Level::DEBUG,
-        _ => tracing::Level::TRACE,
-    };
-    let _ = tracing_subscriber::fmt()
-        .with_max_level(level)
-        .with_target(false)
-        // Logs MUST go to stderr; stdout carries the JSON envelope
-        // that scripts/tests parse. Without this, every `INFO` line
-        // from `JsonlStorage::reload` corrupts the response.
-        .with_writer(std::io::stderr)
-        .try_init();
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use clap::Parser;
-
-    /// The orphan-volume guard refuses a bulk delete, and that refusal is
-    /// only defensible while the user has a way to say the deletion was
-    /// meant. `OrphanGuard::Disabled` is that way, and this binary is the
-    /// only thing that reaches it — so the flag existing is what keeps the
-    /// guard from being a wall (root `CLAUDE.md` invariant 9).
-    ///
-    /// The escape hatch was unreachable from any user-facing surface for
-    /// one commit, which is exactly the state this pins against.
-    #[test]
-    fn the_bulk_delete_escape_hatch_is_reachable_from_the_command_line() {
-        let cli =
-            Cli::try_parse_from(["outl", "reconcile", "--ahead-of-log", "--allow-bulk-delete"])
-                .expect("`--allow-bulk-delete` must parse");
-        assert!(matches!(
-            cli.command,
-            Some(Command::Reconcile {
-                allow_bulk_delete: true,
-                ahead_of_log: true,
-                ..
-            })
-        ));
-    }
-
-    /// And it is off unless asked for.
-    #[test]
-    fn the_guard_is_enforced_by_default() {
-        let cli = Cli::try_parse_from(["outl", "reconcile"]).expect("plain reconcile must parse");
-        assert!(matches!(
-            cli.command,
-            Some(Command::Reconcile {
-                allow_bulk_delete: false,
-                ..
-            })
-        ));
-    }
 }
